Removed unrequired JS and CSS
Bootstrap was used before as easily implementable temporary css so that the pages are easy to comprehend in development. When it was replaced by tailwind, the js and css of bootstrap was still present, but isnt required anymore. Thus it is being removed. Aditionally, there is a script.js which doesn't look required for any functionality, thus its being removed.